feat(workflow): migrate end, note and var-assign to the catalog
Second catalog batch, following the wait template: the data half of each
type moves to catalog/nodes/{end,note,var-assign}.ts; each web schema.ts
shrinks to its defineFromManifest merge site plus back-compat re-exports;
zero diff under any node.tsx / panel.tsx / trace-renderer.tsx.

Along the way:
- extractVarIdsFromString joins catalog/variable-inference.ts (it was
  already pure); the tiptap-converters module re-exports it.
- Engine sweep deletes two shadow declarations: var-assign-processor's local
  VariableAssignment (narrower than the builder's) now imports the catalog
  type, and end.ts's EndNodeConfig becomes a Pick of the catalog data.
- The defaults-parse test caught its first real drift: var-assign's default
  starter row (empty name) failed its own schema's min(1) + identifier
  regex - the resource-trigger bug class. Fixed in the schema, not the
  defaults: an empty name is a legitimate persisted state (half-configured
  nodes save; today's save path never parses), so the schema now allows
  empty-or-valid-identifier while validateVarAssign keeps flagging empties
  as checklist errors.
- var-assign's deprecated, consumer-less varAssignConfigSchema died in the
  move; its default-row id is now minted with generateId instead of uuid.
