+++ apps/web/src/components/workflow/nodes/core/end/schema.ts
@@ -1,69 +1,19 @@
 // apps/web/src/components/workflow/nodes/core/end/schema.ts
 
-import { z } from 'zod'
-import { baseNodeDataSchema } from '~/components/workflow/types/node-base'
-import { extractVarIdsFromString } from '~/components/workflow/ui/input-editor/tiptap-converters'
+import { endManifest, type NodeManifest } from '@auxx/lib/workflow-engine/client'
 import { createUnifiedOutputVariable } from '~/components/workflow/utils/variable-conversion'
-import { BaseType, NodeCategory, type NodeDefinition, type ValidationResult } from '../../../types'
-import { NodeType } from '../../../types/node-types'
+import { BaseType, type NodeDefinition } from '../../../types'
+import { defineFromManifest } from '../../define-from-manifest'
 import type { EndNodeData } from './types'
 
-/**
- * Main schema for end node data (simplified structure)
- */
-export const endNodeDataSchema = baseNodeDataSchema.extend({
-  title: z.string().default('End'),
-  description: z.string().optional(),
-  message: z.string().optional(),
-  status: z.enum(['success', 'error']).optional(),
-})
-
-/**
- * Default configuration for the End node
- */
-export const endDefaultData: Partial<EndNodeData> = {
-  title: 'Output',
-  description: '',
-  message: '',
-  status: 'success',
-}
-
-/**
- * Extract variables from end node message field
- */
-export function extractEndVariables(data: Partial<EndNodeData>): string[] {
-  const variableIds = new Set<string>()
-
-  if (data.message && typeof data.message === 'string') {
-    extractVarIdsFromString(data.message).forEach((id) => variableIds.add(id))
-  }
-
-  return Array.from(variableIds)
-}
-
-/**
- * Validates the End node configuration
- */
-export const validateEndConfig = (data: EndNodeData): ValidationResult => {
-  const errors: Array<{ field: string; message: string; type?: 'warning' | 'error' }> = []
-
-  // Validate title
-  if (!data.title?.trim()) {
-    errors.push({ field: 'title', message: 'Title is required', type: 'error' })
-  }
-
-  // Validate status if provided
-  if (data.status && !['success', 'error'].includes(data.status)) {
-    errors.push({ field: 'status', message: 'Invalid status value', type: 'error' })
-  }
-
-  return { isValid: errors.filter((e) => e.type === 'error').length === 0, errors }
-}
+// The data half (schema, defaults, validator, variable extraction) lives in
+// the node catalog (`@auxx/lib/workflow-engine/catalog/nodes/end`). This file
+// is the merge site: manifest + the web-only output resolver.
 
 /**
  * Define output variables for the end/output node
  */
-const getEndOutputVariables = (data: Partial<EndNodeData>, nodeId: string) => {
+const getEndOutputVariables = (_data: Partial<EndNodeData>, nodeId: string) => {
   return [
     createUnifiedOutputVariable({
       nodeId,
@@ -83,17 +33,19 @@ const getEndOutputVariables = (data: Partial<EndNodeData>, nodeId: string) => {
 /**
  * Node definition for the End node
  */
-export const endDefinition: NodeDefinition<EndNodeData> = {
-  id: NodeType.END,
-  category: NodeCategory.ACTION,
-  displayName: 'Output',
-  description: 'Outputs a message for the manual trigger',
-  icon: 'message-circle',
-  color: '#10b981', // ACTION category color
-  defaultData: endDefaultData,
-  schema: endNodeDataSchema,
-  validator: validateEndConfig,
-  canRunSingle: true,
-  extractVariables: extractEndVariables,
-  outputVariables: getEndOutputVariables as any,
-}
+export const endDefinition: NodeDefinition<EndNodeData> = defineFromManifest(
+  endManifest as unknown as NodeManifest<EndNodeData>,
+  { outputVariables: getEndOutputVariables as any }
+)
+
+// Back-compat re-exports so no consumer import churns:
+export {
+  endNodeDataSchema,
+  extractEndVariables,
+  validateEndConfig,
+} from '@auxx/lib/workflow-engine/client'
+
+/**
+ * Default configuration for the End node
+ */
+export const endDefaultData = endManifest.defaultData() as Partial<EndNodeData>
