+++ apps/web/src/components/workflow/nodes/core/end/types.ts
@@ -1,15 +1,19 @@
 // apps/web/src/components/workflow/nodes/core/end/types.ts
 
-import type { BaseNodeData, SpecificNode } from '~/components/workflow/types'
+import type { CatalogEndNodeData } from '@auxx/lib/workflow-engine/client'
+import type { SpecificNode } from '~/components/workflow/types'
+import type { NodeType } from '~/components/workflow/types/node-types'
+
+// The data half moved to the node catalog (node-catalog Phase 1 —
+// `@auxx/lib/workflow-engine/catalog/nodes/end`). EndNodeData narrows `type`
+// to the web NodeType enum, same as BaseNodeData does over its lib
+// counterpart.
 
 /**
  * Node data structure for the End node with minimal configuration
  */
-export interface EndNodeData extends BaseNodeData {
-  /** Optional message to display when workflow ends */
-  message?: string
-  /** Optional status to set when workflow ends */
-  status?: 'success' | 'error'
+export interface EndNodeData extends CatalogEndNodeData {
+  type: NodeType
 }
 
 /**
