+++ apps/web/src/components/workflow/nodes/core/var-assign/types.ts
@@ -1,36 +1,18 @@
 // apps/web/src/components/workflow/nodes/core/var-assign/types.ts
 
-import type { BaseNodeData, SpecificNode } from '~/components/workflow/types/node-base'
-import type { BaseType } from '~/components/workflow/types/unified-types'
+import type { CatalogVarAssignNodeData } from '@auxx/lib/workflow-engine/client'
+import type { SpecificNode } from '~/components/workflow/types/node-base'
+import type { NodeType } from '~/components/workflow/types/node-types'
 
-/**
- * Single variable assignment configuration
- */
-export interface VariableAssignment {
-  /** Unique identifier for the assignment */
-  id: string
-  /** Variable name */
-  name: string
-  /** Variable type (e.g., STRING, NUMBER, BOOLEAN) */
-  type: BaseType
-  /** Whether this is an array of the specified type */
-  isArray?: boolean
-  /** Variable value(s) - can be string or array of strings for array type */
-  value: string | string[]
-  /** Whether this variable is in constant mode (for single values only) */
-  isConstantMode?: boolean
-  /** Constant mode tracking for each array item (for array values only) */
-  itemConstantModes?: boolean[]
-}
+// The data half moved to the node catalog (node-catalog Phase 1 —
+// `@auxx/lib/workflow-engine/catalog/nodes/var-assign`).
+export type { VariableAssignment } from '@auxx/lib/workflow-engine/client'
 
 /**
  * Variable assignment node data (flattened structure)
  */
-export interface VarAssignNodeData extends BaseNodeData {
-  /** Array of variable assignments */
-  variables: VariableAssignment[]
-  /** Whether to ignore type errors during execution */
-  ignoreTypeError?: boolean
+export interface VarAssignNodeData extends CatalogVarAssignNodeData {
+  type: NodeType
 }
 
 /**
